Compatibility with multifactor authentication of rdb 2.6
When connecting to database, presence of flags isc_dpb_trusted_auth and isc_dpb_multi_factor_auth is checked. If flags are set, the AuthSspi object is initialized and specified authentication factors are performed. As long as server sends responses with the type of op_trusted_auth operation, AuthSspi processes them.
I'm not yet sure of correct location of code for multifactor authentication, perhaps it will have to be moved in future.