Backported ability to exclude plugins using crypto library
Authentication plugins for certificate, password and multifactor plugin use a cryptographic library that is initialized separately. If the cryptographic library is not found in current configuration, then the plugins will throw an exception that will break authentication with other plugins. To correct this problem, it's necessary to check the presence of a cryptographic library and, in case of its absence, transfer to the driver information about excluding these plugins from the list of available for connection.