See #42832. Ability to control statements execution
The added abstract class allows implementing in client application side a mechanism for centrally controlling the execution of statements, as well as their cancellation if the operation takes too much time.

If the class implementation instance in the `StatementOperationAware` is not initialized, then the execution of the statements runs as usual and does not interrupt work.